+++ src/main/java/de/alpharogroup/db/entity/deletable/Deletion.java
@@ -22,25 +22,38 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.db.entity.traceable;
+package de.alpharogroup.db.entity.deletable;
 
 import java.io.Serializable;
 
-import javax.persistence.MappedSuperclass;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 
 import de.alpharogroup.db.entity.base.SequenceBaseEntity;
-import de.alpharogroup.db.entity.creatable.Creation;
-import de.alpharogroup.db.entity.deletable.Deletion;
-import de.alpharogroup.db.entity.modifiable.LastModification;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 
 /**
- * The entity class {@link TraceableEntity} is keeping the information for the creation, last
- * modification and for the deletion of an entity. <br>
+ * The entity class {@link Deletion} acts like an deletion flag that keeps the information for the
+ * deletion of another entity that have this entity as an instance variable. This information are
+ * the date when it was deleted and from whom it was deleted. So this entity can be attached to
+ * another entity for keeping information if it is deleted and when it was deleted. <br>
+ * <br>
+ * In specific situation entities have to be kept for instance of historical purposes. So the intent
+ * of this entity is that a parent entity will not really be deleted. The initial value of the
+ * parent entity that keeps this entity as an instance variable as a null object that signals that
+ * it is not deleted, and when it get a value with the deletion data it signals that it is now
+ * deleted and is not visible to selections. <br>
+ * <br>
+ * Note: So this entity will only be created if the parent entity references to this entity and is
+ * intended to be deleted. This can be done for instance in a deletion strategy.
+ *
+ * Note: There is no need of a delete flag in this entity because this entity is the flag. So if the
+ * parent entity is not any more 'deleted' the reference of this entity will be set to null back.
  *
  * @param <PK>
  *            the generic type of the id
@@ -48,40 +61,27 @@ import lombok.experimental.SuperBuilder;
  *            the generic type of time measurement
  * @param <U>
  *            the generic type of the user or account
- * @see Creation
- * @see Deletion
- * @see LastModification
  */
-@MappedSuperclass
+@Entity
+@Table(name = "deletion")
 @Getter
 @Setter
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
 @SuperBuilder
-public abstract class TraceableEntity<PK extends Serializable, T, U> extends SequenceBaseEntity<PK>
+public class Deletion<PK extends Serializable, T, U> extends SequenceBaseEntity<PK>
 	implements
-		IdentifiableTraceable<PK, T, U>
+		IdentifiableDeletable<PK, T, U>
 {
 
 	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;
-
-	/** The date and time when the entity that owns this entity was created. */
-	private T created;
-
-	/** The user or account that created the entity that owns this entity. */
-	private U createdBy;
 
 	/** The date and time when the entity that owns this entity was deleted. */
 	private T deleted;
 
 	/** The user or account that deleted the entity that owns this entity. */
 	private U deletedBy;
-
-	/** The date and time when the entity that owns this entity was modified. */
-	private T lastModified;
-
-	/** The user or account that modified the entity that owns this entity. */
-	private U lastModifiedBy;
 
 }