+++ src/main/java/de/alpharogroup/db/entity/verifiable/BasicVerifiableEntity.java
@@ -22,23 +22,24 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.db.entity.visibility;
+package de.alpharogroup.db.entity.verifiable;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
 
+import de.alpharogroup.db.entity.Identifiable;
 import de.alpharogroup.db.entity.base.SequenceBaseEntity;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
 /**
- * The class {@link VisibilityEntity} is a base entity with a flag 'visible' that indicates if the
- * entity is visible or not.
+ * The class {@link BasicVerifiableEntity} is a base entity that have to be signed and verified
  *
  * @param <PK>
  *            the generic type of the id
@@ -48,16 +49,18 @@ import lombok.experimental.SuperBuilder;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
-public abstract class VisibilityEntity<PK extends Serializable> extends SequenceBaseEntity<PK>
+public abstract class BasicVerifiableEntity<PK extends Serializable> extends SequenceBaseEntity<PK>
 	implements
-		IdentifiableVisibility<PK>
+		Identifiable<PK>,
+		Verifiable
 {
 
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;
 
-	/** The attribute visible, if true this entity is visible. */
-	@Column(name = "visible")
-	private boolean visible;
+	/** The signature. */
+	String signature;
+
 }